securedrop-admin: move to the standalone admin directory
