+++ admin/bin/validate-gpg-key.sh
