+++ admin/securedrop_admin/__init__.py
@@ -1,4 +1,21 @@
-#!/usr/bin/env python2
+# -*- mode: python; coding: utf-8 -*-
+#
+# Copyright (C) 2013-2018 Freedom of the Press Foundation & al
+# Copyright (C) 2018 Loic Dachary <loic@dachary.org>
+#
+#    This program is free software: you can redistribute it and/or modify
+#    it under the terms of the GNU General Public License as published by
+#    the Free Software Foundation, either version 3 of the License, or
+#    (at your option) any later version.
+#
+#    This program is distributed in the hope that it will be useful,
+#    but WITHOUT ANY WARRANTY; without even the implied warranty of
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#    GNU General Public License for more details.
+#
+#    You should have received a copy of the GNU General Public License
+#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#
 """
 SecureDrop Admin Toolkit.
 
@@ -19,7 +36,7 @@ from prompt_toolkit.validation import Validator, ValidationError
 
 sdlog = logging.getLogger(__name__)
 
-SD_DIR = os.path.realpath(os.path.join(os.path.dirname(__file__), '..'))
+SD_DIR = os.path.realpath(os.path.join(os.path.dirname(__file__), '../../..'))
 ANSIBLE_PATH = os.path.join(SD_DIR, "./install_files/ansible-base")
 APP_PATH = os.path.join(SD_DIR, "securedrop")
 SITE_CONFIG = os.path.join(ANSIBLE_PATH, "group_vars/all/site-specific")
@@ -386,131 +403,6 @@ def sdconfig(args):
     SiteConfig(args).load_and_update_config()
 
 
-def run_command(command):
-    """
-    Wrapper function to display stdout for running command,
-    similar to how shelling out in a Bash script displays rolling output.
-
-    Yields a list of the stdout from the `command`, and raises a
-    CalledProcessError if `command` returns non-zero.
-    """
-    popen = subprocess.Popen(command,
-                             stdout=subprocess.PIPE,
-                             stderr=subprocess.STDOUT)
-    for stdout_line in iter(popen.stdout.readline, ""):
-        yield stdout_line
-    popen.stdout.close()
-    return_code = popen.wait()
-    if return_code:
-        raise subprocess.CalledProcessError(return_code, command)
-
-
-def install_apt_dependencies(args):
-    """
-    Install apt dependencies in Tails. In order to install Ansible in
-    a virtualenv, first there are a number of Python prerequisites.
-    """
-    sdlog.info("Installing SecureDrop Admin dependencies")
-    sdlog.info(("You'll be prompted for the temporary Tails admin password,"
-                " which was set on Tails login screen"))
-
-    apt_command = ['sudo', 'su', '-c',
-                   "apt-get update && \
-                   apt-get -q -o=Dpkg::Use-Pty=0 install -y \
-                   python-virtualenv \
-                   python-pip \
-                   ccontrol \
-                   virtualenv \
-                   libffi-dev \
-                   libssl-dev \
-                   libpython2.7-dev",
-                   ]
-
-    try:
-        # Print command results in real-time, to keep Admin apprised
-        # of progress during long-running command.
-        for output_line in run_command(apt_command):
-            print(output_line.rstrip())
-    except subprocess.CalledProcessError:
-        # Tails supports apt persistence, which was used by SecureDrop
-        # under Tails 2.x. If updates are being applied, don't try to pile
-        # on with more apt requests.
-        sdlog.error(("Failed to install apt dependencies. Check network"
-                     " connection and try again."))
-        raise
-
-
-def envsetup(args):
-    """Installs Admin tooling required for managing SecureDrop. Specifically:
-
-        * updates apt-cache
-        * installs apt packages for Python virtualenv
-        * creates virtualenv
-        * installs pip packages inside virtualenv
-
-    The virtualenv is created within the Persistence volume in Tails, so that
-    Ansible is available to the Admin on subsequent boots without requiring
-    installation of packages again.
-    """
-    # virtualenv doesnt exist? Install dependencies and create
-    if not os.path.exists(VENV_ACTIVATION):
-
-        install_apt_dependencies(args)
-
-        # Technically you can create a virtualenv from within python
-        # but pip can only be run over tor on tails, and debugging that
-        # along with instaling a third-party dependency is not worth
-        # the effort here.
-        sdlog.info("Setting up virtualenv")
-        try:
-            sdlog.debug(subprocess.check_output(['torify', 'virtualenv',
-                                                 VENV_DIR],
-                                                stderr=subprocess.STDOUT))
-        except subprocess.CalledProcessError as e:
-            sdlog.debug(e.output)
-            sdlog.error(("Unable to create virtualenv. Check network settings"
-                         " and try again."))
-            raise
-    else:
-        sdlog.info("Virtualenv already exists, not creating")
-
-    install_pip_dependencies(args)
-
-    sdlog.info("Finished installing SecureDrop dependencies")
-
-
-def install_pip_dependencies(args, pip_install_cmd=[
-        'torify',
-        os.path.join(VENV_DIR, 'bin', 'pip'),
-        'install',
-        # Specify requirements file.
-        '-r', os.path.join(SD_DIR, 'securedrop',
-                           'requirements', 'admin-requirements.txt'),
-        '--require-hashes',
-        # Make sure to upgrade packages only if necessary.
-        '-U', '--upgrade-strategy', 'only-if-needed',
-]):
-    """
-    Install Python dependencies via pip into virtualenv.
-    """
-
-    sdlog.info("Checking Python dependencies for securedrop-admin")
-    try:
-        pip_output = subprocess.check_output(pip_install_cmd,
-                                             stderr=subprocess.STDOUT)
-    except subprocess.CalledProcessError as e:
-        sdlog.debug(e.output)
-        sdlog.error(("Failed to install pip dependencies. Check network"
-                     " connection and try again."))
-        raise
-
-    sdlog.debug(pip_output)
-    if "Successfully installed" in pip_output:
-        sdlog.info("Python dependencies for securedrop-admin upgraded")
-    else:
-        sdlog.info("Python dependencies for securedrop-admin are up-to-date")
-
-
 def install_securedrop(args):
     """Install/Update SecureDrop"""
     activate_venv(args)
@@ -608,9 +500,6 @@ def parse_argv(argv):
     parser.add_argument('--app-path', default=APP_PATH,
                         help="path to the SecureDrop application root")
     subparsers = parser.add_subparsers()
-
-    parse_setup = subparsers.add_parser('setup', help=envsetup.__doc__)
-    parse_setup.set_defaults(func=envsetup)
 
     parse_sdconfig = subparsers.add_parser('sdconfig', help=sdconfig.__doc__)
     parse_sdconfig.set_defaults(func=sdconfig)
