+++ securedrop/Dockerfile
@@ -6,7 +6,7 @@ ARG USER_ID
 ENV USER_ID ${USER_ID:-0}
 
 RUN apt-get update && \
-    apt-get install -y python-pip python-yaml libpython2.7-dev libssl-dev secure-delete \
+    apt-get install -y python-pip libpython2.7-dev libssl-dev secure-delete \
                        gnupg2 ruby redis-server firefox git xvfb haveged curl \
                        gettext paxctl x11vnc enchant
