+++ securedrop/requirements/securedrop-app-code-requirements.in
@@ -7,7 +7,6 @@ Flask
 gnupg
 Jinja2
 jsmin
-prompt_toolkit
 psutil
 pycryptodomex
 pyotp
